requirements.txt
iso639-lang==2.1.0
babel==2.11.0
